Added swagger, updated functionality and other changes
