+++ swagger_docs/get_add_song_to_playlist.yml
@@ -1,0 +1,37 @@
+Retrieve playlist details after adding a song:
+  tags:
+    - Playlists
+  parameters:
+    - name: playlist_id
+      in: path
+      description: The ID of the playlist to which the song will be added
+      required: true
+      type: integer
+    - name: song_id
+      in: path
+      description: The ID of the song to add to the playlist
+      required: true
+      type: integer
+  produces:
+    - application/json
+  responses:
+    200:
+      description: Playlist details with the newly added song
+      schema:
+        type: object
+        properties:
+          PlaylistID:
+            type: integer
+            description: Unique identifier of the playlist
+          Name:
+            type: string
+            description: Name of the playlist
+          Description:
+            type: string
+            description: Description of the playlist
+          Songs:
+            type: array
+            items:
+              $ref: '#/definitions/Song'
+    404:
+      description: Playlist or song not found
