+++ swagger_docs/get_playlist.yml
@@ -1,0 +1,16 @@
+Retrieve a specific playlist by its ID:
+  tags:
+    - Playlists
+  parameters:
+    - in: path
+      name: playlist_id
+      type: integer
+      required: true
+      description: Unique ID of the playlist
+  responses:
+    200:
+      description: Details of the specified playlist
+      schema:
+        $ref: '#/definitions/Playlist'
+    404:
+      description: Playlist not found
